[findmaxima]: (4) move code into library/functions
