+++ src/subpackages/library/cellprofiler_library/modules/_findmaxima.py
@@ -1,11 +1,10 @@
 import numpy
 import scipy.ndimage
-from skimage.feature import peak_local_max
 from typing import Optional, Annotated, Union
 from pydantic import validate_call, ConfigDict, Field
 from cellprofiler_library.types import ImageAny, ObjectSegmentation, ImageBinary
 from cellprofiler_library.opts.findmaxima import BackgroundExclusionMode
-
+from cellprofiler_library.functions.image_processing import find_maxima_in_data, find_maxima_threshold, apply_target_mask
 
 @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
 def find_maxima(
@@ -30,33 +29,4 @@ def find_maxima(
         raise NotImplementedError("Invalid background method choice")
     
     return find_maxima_in_data(x_data, min_distance_value, label_maxima, th_abs)
-
-
-def find_maxima_in_data(
-        x_data: ImageAny, 
-        min_distance_value: int, 
-        label_maxima: bool, 
-        th_abs: Optional[float]=None
-    ) -> Union[ImageBinary, ObjectSegmentation]:
-    maxima_coords = peak_local_max(
-        x_data,
-        min_distance=min_distance_value,
-        threshold_abs=th_abs,
-    )
-    y_data = numpy.zeros(x_data.shape, dtype=bool)
-    y_data[tuple(maxima_coords.T)] = True
-
-    if label_maxima:
-        y_data = scipy.ndimage.label(y_data)[0]
-    return y_data
-
-def find_maxima_threshold(x_data: ImageAny) -> ImageAny:
-    """Identity function, just returns the input"""
-    return x_data
-
-def apply_target_mask(x_data: ImageAny, masking_image: Union[ImageAny,ObjectSegmentation]) -> ImageAny:
-    mask = masking_image.astype(bool)
-    x_data[~mask] = 0
-    return x_data
-
